backend/requirements.txt
flask
pymongo[srv]
cerberus
pyyaml
requests